requirements.txt
Flask
python-dotenv
tweepy
PyYAML
jinja2